+++ website/docs/table-group/data-table/examples/table-in-table-with-fixed.jsx
@@ -1,15 +1,14 @@
 import React, { useState } from 'react';
-import { scaleLinear } from 'd3-scale';
 import DataTable from '@semcore/data-table';
 import Accordion from '@semcore/accordion';
 import { Flex } from '@semcore/flex-box';
-import { Plot, Line, XAxis, YAxis, ResponsiveContainer, minMax } from '@semcore/d3-chart';
 
 const RowAccordion = React.forwardRef(function ({ value, collapse = {}, ...props }, ref) {
   return (
     <Accordion.Item value={value} ref={ref}>
       <Accordion.Item.Toggle {...props} />
-      <Accordion.Item.Collapse {...collapse} />
+      {/* [3] Set a variable to remove overflow */}
+      <Accordion.Item.Collapse {...collapse} overflowHidden={false} />
     </Accordion.Item>
   );
 });
@@ -17,27 +16,34 @@ const RowAccordion = React.forwardRef(function ({ value, collapse = {}, ...props
 export default () => {
   const [value, setValue] = useState([]);
   return (
-    /* [1] Wrapping the table in the Accordion control component; */
     <Accordion value={value} onChange={setValue}>
       <DataTable data={data}>
-        <DataTable.Head>
-          <DataTable.Column name="keyword" children="Keyword" />
+        <DataTable.Head wMin={1000}>
+          <DataTable.Column name="keyword" children="Keyword" fixed="left" />
           <DataTable.Column name="kd" children="KD,%" />
           <DataTable.Column name="cpc" children="CPC" />
           <DataTable.Column name="vol" children="Vol." />
         </DataTable.Head>
         <DataTable.Body>
-          {/* [2] Replacing the tag in DataTable.Row with our extended tag with Accordion.Item */}
           <DataTable.Row tag={RowAccordion}>
             {(props, row, index) => {
               return {
-                /* [3] Setting the value for Accordion.Item; */
                 value: index,
-                /* [4] Calculating the active line to highlight it */
                 active: value.includes(index),
                 collapse: {
-                  /* [5] Render the children to accordion content; */
-                  children: <ChartExample />,
+                  children: (
+                    <DataTable data={data}>
+                      {/* [1] Set the desired z-index */}
+                      <DataTable.Head hidden z-index={1}>
+                        <DataTable.Column name="keyword" flex="inherit" fixed="left" />
+                        <DataTable.Column name="kd" flex="inherit" />
+                        <DataTable.Column name="cpc" flex="inherit" />
+                        <DataTable.Column name="vol" flex="inherit" />
+                      </DataTable.Head>
+                      {/* [2] Set a variable to block the scroll */}
+                      <DataTable.Body disabledScroll />
+                    </DataTable>
+                  ),
                 },
               };
             }}
@@ -47,7 +53,6 @@ export default () => {
               return {
                 children: (
                   <Flex alignItems="center">
-                    {/* [6] Set the arrow (Chevron icon), if necessary. */}
                     <Accordion.Item.Chevron color="stone" mr={2} />
                     {props.children}
                   </Flex>
@@ -58,39 +63,6 @@ export default () => {
         </DataTable.Body>
       </DataTable>
     </Accordion>
-  );
-};
-
-const ChartExample = () => {
-  const [[width, height], updateSize] = useState([0, 0]);
-  const MARGIN = 40;
-  const dataChart = Array(20)
-    .fill({})
-    .map((d, i) => ({
-      x: i,
-      y: Math.random() * 10,
-    }));
-  const xScale = scaleLinear()
-    .range([MARGIN, width - MARGIN])
-    .domain(minMax(dataChart, 'x'));
-  const yScale = scaleLinear()
-    .range([height - MARGIN, MARGIN])
-    .domain([0, 10]);
-  return (
-    <ResponsiveContainer h={300} onResize={updateSize}>
-      <Plot data={dataChart} scale={[xScale, yScale]} width={width} height={height}>
-        <YAxis>
-          <YAxis.Ticks />
-          <YAxis.Grid />
-        </YAxis>
-        <XAxis>
-          <XAxis.Ticks />
-        </XAxis>
-        <Line x="x" y="y">
-          <Line.Dots display />
-        </Line>
-      </Plot>
-    </ResponsiveContainer>
   );
 };